feat: introduce new repository interfaces and mappers for Brand, SubCategory, Specification, and SpecificationValue, enhancing data management and retrieval capabilities in the application

+++ server/src/main/java/com/accomputers/api/application/dtos/createProductDTO.java
@@ -7,11 +7,11 @@ import org.springframework.web.multipart.MultipartFile;
 public record createProductDTO(
                 String name,
                 String description,
-                double price,
+                Float price,
                 String condition,
-                double discount,
-                int brandId,
-                int subCategoryId,
+                Float discount,
+                Integer brandId,
+                Integer subCategoryId,
                 List<ImageDTO> images,
                 List<ProductSpecificationDTO> specifications) {
